recast using u64 than usize

+++ backend/crates/server/src/fhir_client/middleware/storage.rs
@@ -233,7 +233,7 @@ impl<
                             )
                             .await?;
 
-                        if search_results.entries.len() > delete_limit as usize {
+                        if search_results.entries.len() as u64 > delete_limit {
                             return Err(OperationOutcomeError::error(
                                 IssueType::invalid(),
                                 format!(
